Proof-of-concept from Queer Lunch and Code
We (@46bit, Alice and Zac) came up with the concept of controlling a
Pong cursor by whether a phone was plugged in (down) or not (up),
found that `usblib` can't detect iOS devices but that you can use a
(slow) OS X shell command `system_profiler -xml SPUSBDataType` which
does, and adapted an existing Pong game we found on Github Gists to
have two players controlled by two different phones. Right now it is
hardcoded such that player one is my phone.

`ioreg -p IOUSB` is a far faster command which can detect iOS devices
but it does not output identifying information.